requirements.txt
streamlit
python-docx
openai
python-dotenv
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
tiktoken
numpy
scikit-learn
git